backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
minio==7.2.7
kafka-python==2.0.2
wandb==0.17.3
python-dotenv==1.0.1
requests==2.32.3
numpy==1.26.4
opencv-python-headless==4.9.0.80
mediapipe==0.10.14
torch==2.3.0
torchvision==0.18.0
transformers==4.41.2
hdbscan>=0.8.33,<0.9
umap-learn==0.5.6
scikit-learn==1.4.2
scipy==1.13.1
matplotlib==3.9.0
Pillow==10.3.0
